changed UP state
there were problems with setting state from UP to ALIVE in Node.move(), because Node.move is called in thread different from thread in which Algorithm is running. So there were times when Node was already at root position, but it was still in UP state so there was nothing on the screen.
(and I don't know how to use English tenses... so hope last sentence makes sence)

This problem could be solved in this way, but there is a need to change it also in other data structures