we are almost ready to create pull request
Scenario (previous button and save button) is defaultly disabled. It's enabled only for AATree, AVLTree and BSTTree, which works well with scenario.
So we can merge it to master (with no problems with compatibility with other DSs) and scenario for other DSs could be added in future.
Motivation - create smaller and more frequent pull requests with new functionality.